Revert to using Custom.Windows package version 1.0.2.1
Custom.Windows usage was updated to version 1.1.0.0 in #64940bb1f0a9ca98e19ac32bacb769d2081dcd67 but there was a problem with the ```IsFirstInstance``` property's value returning null in the ```OnExit``` handler in ```MefBootstrapper.cs```. This caused the application not to run the code saving the preferences file in the ```OnExit``` handler.

The current commit reverts this update. A future commit with the correct change in ```MefBootstrapper.cs``` will handle the correct use of ```IsFirstInstance```.