New private method SetDebugApiEndpoint().
This method is debug only and is used to set a custom endpoint for the DF API. The endpoint variable has to be set in the DebugServerBaseAddress application user setting from the Project's Settings designer. This setting takes precedence over the old way of changing the api endpoint through the use of the endpoint.txt in the application's local app data folder.