Fix error handling when there are exceptions because of expired tokens, fixes #78.
This use case was correctly handled while the app was running, but with two bad side effects. First, the deletion of the upload crashed the app because of an unhandled AmazonS3Exception thrown when calling the abortion of a not yet completed multipart file upload (if that was the pending file upload when the token expired). Second, after restarting the app, the problematic upload was marked for automatic resume, but the same exception was thrown after trying to calculate the total uploaded size of the upload operation. The exception was correctly handled, but the operation status was left as pending instead of paused as it should, so the buttons were invisible. Both are fixed now.