When creating a new upload, set the LocalUpload.Url value to the Archive.Url value used to create the upload.
This was done only after successfully creating the new upload and used the Upload.Url value, which is of course equal to the Archive.Url value. If for any reason the creation failed, then the LocalUpload.Url was left empty. That resulted in errors while loading the uploads after client start.