Add support for getting parent process.
When the instance is not the first one running, then check if parent process is an older instance of the app and wait until this is exited before going on. If the parent process is not an instance of the bigstash app, then shutdown since we want the actual first instance to keep running.