Changes to support settings migration and graceful restarts.
- Don't check for setting SettingsUpgradeRequired when handling OnStartup. Just call the new TryMigratingOldUserConfig method.
- OnExit handler checks for setting RestartAfterUpdate and updates its value to false. If restart is requested, then call SquirrelHelper.RunUpdatedExe() to run the new updated exe after shutting down. Also, if SettingsUpgradeRequired is true, then the current assembly's user.config is copied for future migration.

New method implementation TryMigratingOldUserConfig. This updates SettingsUpgradeRequired to false and saves it. This way the user.config is created if this is the first ever execution of the assembly. Next, a check is done for migrate.user.config in the BigStash AppData folder. If it exists it replaces the user.config, so that settings from previous sessions are migrated. Finally, migrate.user.config is deleted and SettingsUpgradeRequired is set to false once again and saved.