Change 'minimized' to '-m' since it's a more appropriate command line argument.
Also, when activating and checking if there's a registry key for running the app at startup, check if it points to the current assembly's location, and if not, update it to that location value.

We need this to update the registry key after running an updated version of the app.