Squirrel should handle custom bigstash registry keys.
Specifically, a key should exist under HKCU\SOFTWARE with the assembly's name. That key should have a LatestVersionInstallPath variable with the latest app exe location as its value.

This key is to be used by the shell context menu extension, so it can execute the latest version of the app.

The above is handled by CreateOrUpdateCustomRegistryEntries (previously CreateCustomRegistryEntries).

Finally, RemoveCustomRegistryEntries takes care of deleting the custom bigstash key.