Implement QueryContextMenu method.
This method is responsible for querying the selected file so it can decide if a menu needs to be added in the shell's right-click context menu.

CMF_DEFAULTONLY and CMF_VERBSONLY are two flags that indicate we shouldn't add anything. The first ones indicates that the action was the default action on the selected file, that is a double click. The second one is for cases when the user clicked a shortcut file. Since QueryContextMenu gets executed more than once since we register the extension under '*' and 'lknfile', it will add nothing for the first case but it will add our option for the second case. We need that because in the first case the selected path isn't the shortcut file's path, but its target's path.